[REVISIT]Kernelflinger: Boot into fastboot/recovery while powering on.
BMDA 1622

This patch enables boot into fastboot/recovery on long pressing
Volume down key while powering on the device.
This patch increases wait time for control reset from 50
to 150 millisecond. In BYT_M we need to wait for a little
longer to detect the key. Marked as revisit as we need to
check the impact on boot time and if can be fixed in BIOS.

Change-Id: I2e2b8e55b9412e93bd0d567d17b19d21e9f82894
Signed-off-by: shaunak saha <shaunak.saha@intel.com>