boot flow: check battery level
The platform boots in CHARGING mode when the battery level is bellow
the threshold defined in the OEM1 ACPI table and a USB charger is
plugged-in.  If no charger is plugged-in, the platform turns off.

In the process, this patch homogenizes the error code returned when an
image resource if not found.

Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-11216
Change-Id: If1fd7c1a760db483ceb41ae2a421d3c498f10b36
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com:443/307431