cmdline: Disable pre-pending serial-port in cmdline
Issue: BMDA-1695, 1631

Disabling pre-pending of serial-port(tty0) in cmdline from
Kernelflinger, as this is causing conflict with the normal
kernel commandline settings from Boardconfig.mk of
device/intel/baytrail/byt_m_crb.

Moreover we do not use the tty0 as the console on Baytrail-M
platform and also this is causing kernel logs come on DUT screen
during boot.

Change-Id: Ibcfa91ef0f20116b5da1c2f1c1daa574dd7a2aae
Signed-off-by: Mantesh sarasetti <manteshx.sarasetti@intel.com>