fastboot: improve the command restriction mechanism
The cmd_register() function takes the minimum device state from which
the command is allowed.  States are defined in the following order :
LOCKED, VERIFIED and UNLOCKED.  Which means that if a command is
allowed in VERIFIED state this command is allowed in UNLOCKED state
also but prohibited in LOCKED state.

Disable the oem setvar and garbage-disk commands in verified state.

Change-Id: I1f11b64d694a04d0647a09cd682aded3dda84f7d
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-15381
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com:443/313781