watchdog fallback policy
In the events of multiple watchdog resets (currently more than 2) in a
short period of time, the bootloader now offers the user the ability
to choose the next option instead of keeping crashing again and again.

This behavior is enabled as the default behavior.  Yet, in order to
not block the automatic tests, it can be disable using the following
fastboot command:

$ fastboot oem crash-event-menu 0

Change-Id: Id081cb3e4fb2bc8f1e377a72562411c760e12cdb
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-15383
Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>
Reviewed-on: https://android.intel.com:443/313785